Changed some expressions
This change makes it more clear, that we added a new table view (cards aren't new). As we don't have any sign-in modal anymore, it's unpracticable to say that the modal is on a new page but rather that the content of the modal has moved. Sentences start now with a capital letter too.