Improved video embedding regexes
Vimeo video-links may contain the channel of the uploader, the collection id and more parameters in front of the video id. This change fixes the regex to recognize such links too. [1]
YouTube links might contain additional query parameters like the playlist or referer. The YouTube regex was updated to support embedding of links with these parameters.

[1]: see the documentation for all url schemes at https://developer.vimeo.com/api/oembed/videos