Change note detail redux state
This now works with NoteDto, as this is what we'll get from the api.
Changed `userId` to `userName` in LastChange, because the backend only gives out usernames and never the id.
Changed `authorship` from an array of numbers to strings as this is what the backend now provides.
Removed `preVersionTwoNote` field, because it isn't used and probably won't be needed.

Signed-off-by: Philip Molares <philip.molares@udo.edu>