Remove Firefox from CI
Okay I'm just done. Firefox in our current CI is just the WORST. It is slower than chrome and crashes A LOT! Like really! The e2e workflows of some PRs have to run 4 or 5 times because firefox decides to just lose the connection.
This may also be caused by the performance of GHA, but right now it's easier to (temporally) remove firefox than switching to a whole new CI.
Just to be clear: I know that this step isn't optimal. I know that a huge part of our community uses firefox. But right now the flakiness of the firefox CI runs drives me crazy!
At the moment we need the e2e tests just for our development to make sure that we don't accidentally break something.
We can re-add firefox again when it's stable to use. Or when we have shifted enough of the e2e tests to jest. It would be nice to test other browsers (like safari or edge) as well.

Signed-off-by: Tilman Vatteroth <git@tilmanvatteroth.de>